Fix ambigiousity in ResourceManager when resolving INI inclusion. ResourceManager takes mod_or_package as an input argument. We want a package, not a module, because module may resolve to an arbitraty namespaced package and is not table.
Do not touch or move INI files around. These INI files are covered in 30+ pages of documentation and process and also referred by other packages. Arbitraty changes for them needs at least 5 workdays worth of work. Leave them as is and do a overhaul of config file system, with a preferably YAML based solution, in Websauna 1.1 or later.